+++ efgs-server-service/src/main/java/es/gob/radarcovid/efgs/runner/CleanBatchJobExecutionRunner.java
@@ -7,37 +7,40 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package es.gob.radarcovid.efgs.batch;
+package es.gob.radarcovid.efgs.runner;
 
 import org.slf4j.MDC;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 import es.gob.radarcovid.efgs.business.CleanBatchJobService;
 import es.gob.radarcovid.efgs.etc.Constants;
-import lombok.RequiredArgsConstructor;
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
 
 @ConditionalOnProperty(name = "application.efgs.clean-batch-job.enabled", havingValue = "true")
 @Component
-@RequiredArgsConstructor
+@Order(3)
+@AllArgsConstructor
 @Slf4j
-public class CleanBatchJobExecutionBatch {
-
+public class CleanBatchJobExecutionRunner extends EfgsRunner {
+	
 	private static final String CLEAN_BATCH_JOB_NAME = "cleanBatchJobExecution";
-
+	
 	private final CleanBatchJobService cleanBatchJobService;
+	
+	@Override
+	public String jobName() {
+		return CLEAN_BATCH_JOB_NAME;
+	}
 
-    @Scheduled(cron = "${application.efgs.clean-batch-job.batching.cron}", zone = "Europe/Madrid")
-	@SchedulerLock(name = CLEAN_BATCH_JOB_NAME, lockAtLeastFor = "PT30S", 
-		lockAtMostFor = "${application.efgs.clean-batch-job.batching.lock-limit}")
-	public void efgsCleanBatchJobScheduledTask() {
+	@Override
+	public void run() {
     	MDC.put(Constants.TRACKING, "CLEAN_BATCH_JOB");
-		log.info("CleanBatchJobBatch process started");
+		log.info("CleanBatchJobExecutionRunner started");
 		cleanBatchJobService.cleanBatchJobExecution();		
-		log.info("CleanBatchJobBatch process finished");
+		log.info("CleanBatchJobExecutionRunner finished");
 	}
 
 }